Added a abstract result class for future resulttypes
